backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Timz"
authors = [
    {name = "Elisa Cohen",email = "elisa.cohen05@gmail.com"}
]
readme = "README.md"
packages = [{ include = "app" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "firebase-admin (>=7.1.0,<8.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]


[tool.isort]
profile = "black"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
pre-commit = "^4.3.0"
asgi-lifespan = "^2.1.0"
pytest-forked = "^1.6.0"

[tool.poetry]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Timz"
package-mode = false
